requirements.txt
selenium>=2.35.0
Pillow>=2.3.0
SimpleCV>=1.3
numpy>=1.8.2